// packages/demo-nextjs/src/app/admin/(cms)/products/core/types.ts
export interface Product {
  id: string;
  priceRange: {
    minVariantPrice: {
      amount: number;
      currencyCode: string;
    };
  };
  description: string;
  title: string;
  totalInventory: number;
  featuredImage?: {
    url: string;
  };
  availableForSale: boolean;
  createdAt: string;
  vendor: string;
  url: string;
}